Clean up AppNavigator.js
