+++ navigation/AppNavigator.js
@@ -1,25 +1,19 @@
 import React from 'react';
-import { AsyncStorage, View, TouchableOpacity, Linking } from 'react-native';
+import { AsyncStorage } from 'react-native';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
-import { createDrawerNavigator, DrawerItems } from 'react-navigation-drawer';
-import { createStackNavigator } from 'react-navigation-stack';
-import WelcomeScreen from '../screens/auth/WelcomeScreen';
-import { Title } from '../components/BaseComponents';
-import LoginScreen from '../screens/auth/LoginScreen';
-import SignUpScreen from '../screens/auth/SignUpScreen';
+import { createDrawerNavigator } from 'react-navigation-drawer';
 import Colors from '../assets/Colors';
-import { getUser } from '../lib/rewardsUtils';
-import { RewardsStack, StoresStack, ResourcesStack } from './StackNavigators';
-
-const AuthStack = createStackNavigator({
-  Welcome: WelcomeScreen,
-  Login: LoginScreen,
-  SignUp: SignUpScreen
-});
+import DrawerContent from './DrawerContent';
+import {
+  AuthStack,
+  ResourcesStack,
+  RewardsStack,
+  StoresStack
+} from './StackNavigators';
 
 class AuthLoadingScreen extends React.Component {
-  constructor() {
-    super();
+  constructor(props) {
+    super(props);
     this._bootstrapAsync();
   }
 
@@ -40,80 +34,6 @@ class AuthLoadingScreen extends React.Component {
   // Render any loading content that you like here
   render() {
     return null;
-  }
-}
-
-export class DrawerContent extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      user: {
-        id: null,
-        name: null
-      },
-      link: 'http://tiny.cc/RewardsFeedback'
-    };
-  }
-  async componentDidMount() {
-    const userId = await AsyncStorage.getItem('userId');
-    getUser(userId).then(userRecord => {
-      if (userRecord) {
-        const user = {
-          id: userId,
-          name: userRecord.fields.Name
-        };
-        this.setState({ user });
-        return true;
-      }
-      return false;
-    });
-  }
-
-  _logout = async () => {
-    AsyncStorage.clear();
-    this.props.navigation.navigate('Auth');
-  };
-
-  render() {
-    return (
-      <View
-        style={{
-          display: 'flex',
-          flex: 1,
-          flexDirection: 'column'
-        }}>
-        <View
-          style={{
-            backgroundColor: Colors.black,
-            height: 114,
-            display: 'flex',
-            flexDirection: 'row',
-            alignItems: 'flex-end',
-            padding: 16
-          }}>
-          <Title style={{ color: 'white' }}>{this.state.user.name}</Title>
-        </View>
-        <DrawerItems {...this.props} />
-        <View
-          style={{
-            flex: 1,
-            flexDirection: 'column',
-            justifyContent: 'flex-end',
-            verticalAlign: 'bottom'
-          }}>
-          <TouchableOpacity
-            style={{ padding: 16 }}
-            onPress={() => Linking.openURL(this.state.link)}>
-            <Title>Report Issue</Title>
-          </TouchableOpacity>
-          <TouchableOpacity
-            style={{ paddingLeft: 16, paddingBottom: 21 }}
-            onPress={() => this._logout()}>
-            <Title>Logout</Title>
-          </TouchableOpacity>
-        </View>
-      </View>
-    );
   }
 }
